backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
numpy==2.1.1
faiss-cpu==1.9.0
pymupdf==1.25.1
sentence-transformers==3.0.1
openai==1.54.4
httpx==0.27.2
python-multipart==0.0.9

